Initial attempt at making a Metlink integration based on the new API.
Previously I used the RESTful integration to pull from the unofficial json URL,
but this has partially stopped working now, and an official API is released.
Because of the way they require the API key to be passed, this can no longer
be handled ad-hoc by the RESTful integration, and requires a custom component.