Handle null values returned by the API better.
None as a dict value is different than it not being set, so using the get()
default does not work - check for None afterwards instead.

Avoid setting the "name" attribute - change it to "service_name".
Setting "name" as an attribute seems to interfere with the name() property.

Set the friendly_name to a combination of the route id and destination name
for the next service.
This saves having to use a template to get a sensible output in the UI.

Bump version to 0.1.0 since the API seems to be working and after these changes should be good to use.