Set the state to a string rather than a datetime.
As reported on issue #2, iOS is fussy about the format of timestamp strings.
The correct ISO format is actually returned by the Metlink API, but when
we parse it into a datetime object, Home Assistant later converts it back
using a more human readable format (a space instead of T between date and time).

Although I had a lot of trouble getting timestamps to display correctly
early on, it seems the format from Metlink API is OK to use as-is.