Merge options with initial config, allow stops to be empty.
When trying to edit an existing config, the following error was thrown:
```
   File "/config/custom_components/metlink/config_flow.py", line 154, in async_step_init
      updated_stops = deepcopy(self.config_entry.data[CONF_STOPS])
   KeyError: 'stops'
```
This was due to (perhaps manual) migration of stops from config.data to
config.options.  This should also resolve a previous issue I had of old
config coming back, which was due to the reconfig only using the list of
stops from initial config.