async_get_registry renamed to async_get.
This rename happened a year ago, but since it is only in the config_flow,
the deprecation warning was not triggered in normal usage.

Now the deprecated function alias has been removed it is an error on import,
which stops the integration from working.

It seems the replacement function may have been introduced in 2021.2,
but the deprecation warning was only added in 2022.6, so require that
version to be safe (it is already very old).