config_flow: async_get is not awaitable
async_get is apparently has the async prefix because it is a callback,
not because it is awaitable.  Not sure why this only stopped working
in HA 2023.6, but there appear to have been some dropped deprecations
of older truly async functions for which this was the replacement, so
it seems some compatibility shim has been removed.

Issue #7